# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/stock_entry_items:
    get:
      tags:
        - StockEntryItems
      summary: List StockEntryItems
      description: Get a paginated list of StockEntryItems with optional filtering and sorting
      operationId: listStockEntryItems
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StockEntryItemList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - StockEntryItems
      summary: Create StockEntryItem
      description: Create a new StockEntryItem
      operationId: createStockEntryItem
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateStockEntryItemRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StockEntryItem'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/stock_items:
    get:
      tags:
        - StockItems
      summary: List StockItems
      description: Get a paginated list of StockItems with optional filtering and sorting
      operationId: listStockItems
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StockItemList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - StockItems
      summary: Create StockItem
      description: Create a new StockItem
      operationId: createStockItem
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateStockItemRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StockItem'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    StockEntryItem:
      type: object
      required:
        - id
        - entry_id
        - company_id
        - item_id
        - quantity
        - metadata
      properties:
        id:
          type: string
          format: uuid
        entry_id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        item_id:
          type: string
          format: uuid
        quantity:
          type: string
        metadata:
          type: object
...
    CreateStockEntryItemRequest:
      type: object
      required:
        - entry_id
        - company_id
        - item_id
        - quantity
        - metadata
      properties:
        entry_id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        item_id:
          type: string
          format: uuid
        quantity:
          type: string
        metadata:
          type: object
...
    StockEntryItemList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/StockEntryItem'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    StockItem:
      type: object
      required:
        - id
        - item_id
        - company_id
        - stock_uom
        - is_stock_item
        - has_batch
        - valuation_method
        - reorder_level
        - metadata
      properties:
        id:
          type: string
          format: uuid
        item_id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        stock_uom:
          type: string
          maxLength: 20
        is_stock_item:
          type: boolean
        has_batch:
          type: boolean
        valuation_method:
          type: string
        reorder_level:
          type: string
        metadata:
          type: object
...
    CreateStockItemRequest:
      type: object
      required:
        - item_id
        - company_id
        - stock_uom
        - is_stock_item
        - has_batch
        - valuation_method
        - reorder_level
        - metadata
      properties:
        item_id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        stock_uom:
          type: string
          maxLength: 20
        is_stock_item:
          type: boolean
        has_batch:
          type: boolean
        valuation_method:
          type: string
        reorder_level:
          type: string
        metadata:
          type: object
...
    StockItemList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/StockItem'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []